% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pepan.R
\name{pepan}
\alias{pepan}
\title{Cumulative distribution function (cdf) for an uncensored epanechnikov distribution}
\usage{
pepan(x = 0, mu = 0, r = 5^0.5)
}
\arguments{
\item{x}{point on x-axis.}

\item{mu}{mean of distribution.}

\item{r}{half the range of the distribution, ie the distance from the mean to the smallest/largest value supported by the distribution. \code{r=5^.5} corresponds to a standard deviation of 1.}
}
\value{
probability of value below \code{x} given \code{mu} and \code{r}.
}
\description{
The inverse of this function is \code{qepan}.
}
\examples{
#Probability of a value below -1.96:
pepan(x=-1.96,mu=0,r=5^.5)

#Cumulative distribution function of epanechnikov distribution:
curve(pepan(x),col='blue',xlim=c(-2.5,2.5),yaxs='i',xaxs='i',
main='Cumulative distribution function',ylab='Probability')

#Cumulative distribution function of standard normal distribution:
curve(pnorm(x),col='green',add=TRUE)

#Legend
legend(x=-2.5,y=1,legend=c('Epanechnikov cdf','Normal cdf'),lty=c(1,1),col=c('blue','green'))
}
\keyword{distribution}

