% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qepan.R
\name{qepan}
\alias{qepan}
\title{Quantile function for an uncensored epanechnikov distribution}
\usage{
qepan(p, mu = 0, r = 5^0.5)
}
\arguments{
\item{p}{probability.}

\item{mu}{mean of distribution.}

\item{r}{half the range of the distribution, ie the distance from the mean to the smallest/largest value supported by the distribution. \code{r=5^.5} corresponds to a standard deviation of 1.}
}
\value{
the quantile associated with \code{x}, \code{mu} and \code{r}.
}
\description{
The inverse of this function is \code{pepan}.
}
\examples{
#Calculate the lower quartile of an epan-distributed variable:
qepan(p=.25,mu=0,r=sqrt(5))

#Use qepan to confirm analytical solution
#Find the quantile corresponding to p=(5+sqrt(5))/8=.9045 when mu=0 and r=sqrt(5):
qepan(p=(5+sqrt(5))/8,mu=0,r=sqrt(5))
#This is equal to
(5-sqrt(5))/2
}
\keyword{distribution}

