% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repan.R
\name{repan}
\alias{repan}
\title{Generate random uncensored epanechnikov-distributed data}
\usage{
repan(n, mu = 0, r = 5^0.5)
}
\arguments{
\item{n}{number of data points.}

\item{mu}{mean of distribution.}

\item{r}{half the range of the distribution, ie the distance from the mean to the smallest/largest value supported by the distribution. \code{r=5^.5} corresponds to a standard deviation of 1.}
}
\value{
vector of random variables.
}
\description{
This function works in conjuncture with \code{qepan} and \code{runif}
}
\examples{
#Generate and plot 10000 random observations:
hist(repan(10000,mu=100,r=10))
}
\keyword{distribution}

